+++ power_loop/contrib/__init__.py
@@ -1,0 +1,4 @@
+"""Optional, dependency-free integrations. Import the module you need:
+
+    from power_loop.contrib.logging_sink import attach_logging_sink
+"""